# Conflicts:
#	composer.json

+++ src/Auth/User.php
@@ -21,11 +21,13 @@ use Statamic\Contracts\Query\ContainsQueryableValues;
 use Statamic\Contracts\Search\Searchable as SearchableContract;
 use Statamic\Data\ContainsComputedData;
 use Statamic\Data\HasAugmentedInstance;
+use Statamic\Data\HasDirtyState;
 use Statamic\Data\TracksQueriedColumns;
 use Statamic\Data\TracksQueriedRelations;
 use Statamic\Events\UserCreated;
 use Statamic\Events\UserCreating;
 use Statamic\Events\UserDeleted;
+use Statamic\Events\UserDeleting;
 use Statamic\Events\UserSaved;
 use Statamic\Events\UserSaving;
 use Statamic\Facades;
@@ -38,7 +40,7 @@ use Statamic\Support\Str;
 
 abstract class User implements Arrayable, ArrayAccess, Augmentable, Authenticatable, AuthorizableContract, CanResetPasswordContract, ContainsQueryableValues, HasLocalePreference, ResolvesValuesContract, SearchableContract, UserContract
 {
-    use Authorizable, CanResetPassword, ContainsComputedData, HasAugmentedInstance, HasAvatar, Notifiable, ResolvesValues, Searchable, TracksQueriedColumns, TracksQueriedRelations;
+    use Authorizable, CanResetPassword, ContainsComputedData, HasAugmentedInstance, HasAvatar, HasDirtyState, Notifiable, ResolvesValues, Searchable, TracksQueriedColumns, TracksQueriedRelations;
 
     protected $afterSaveCallbacks = [];
     protected $withEvents = true;
@@ -140,6 +142,11 @@ abstract class User implements Arrayable, ArrayAccess, Augmentable, Authenticata
         return $this->password();
     }
 
+    public function getAuthPasswordName()
+    {
+        return 'password';
+    }
+
     /**
      * Get or set the blueprint.
      *
@@ -192,11 +199,17 @@ abstract class User implements Arrayable, ArrayAccess, Augmentable, Authenticata
             UserSaved::dispatch($this);
         }
 
+        $this->syncOriginal();
+
         return $this;
     }
 
     public function delete()
     {
+        if (UserDeleting::dispatch($this) === false) {
+            return false;
+        }
+
         Facades\User::delete($this);
 
         UserDeleted::dispatch($this);
