+++ src/Forms/Form.php
@@ -3,6 +3,7 @@
 namespace Statamic\Forms;
 
 use Illuminate\Contracts\Support\Arrayable;
+use Illuminate\Support\Facades\Log;
 use Statamic\Contracts\Data\Augmentable;
 use Statamic\Contracts\Data\Augmented;
 use Statamic\Contracts\Forms\Form as FormContract;
@@ -12,6 +13,7 @@ use Statamic\Events\FormBlueprintFound;
 use Statamic\Events\FormCreated;
 use Statamic\Events\FormCreating;
 use Statamic\Events\FormDeleted;
+use Statamic\Events\FormDeleting;
 use Statamic\Events\FormSaved;
 use Statamic\Events\FormSaving;
 use Statamic\Facades\Blueprint;
@@ -25,6 +27,7 @@ use Statamic\Statamic;
 use Statamic\Support\Arr;
 use Statamic\Support\Str;
 use Statamic\Support\Traits\FluentlyGetsAndSets;
+use Statamic\Yaml\ParseException;
 
 class Form implements Arrayable, Augmentable, FormContract
 {
@@ -241,6 +244,10 @@ class Form implements Arrayable, Augmentable, FormContract
      */
     public function delete()
     {
+        if (FormDeleting::dispatch($this) === false) {
+            return false;
+        }
+
         $this->submissions()->each->delete();
 
         File::delete($this->path());
@@ -313,9 +320,16 @@ class Form implements Arrayable, Augmentable, FormContract
         $path = config('statamic.forms.submissions').'/'.$this->handle();
 
         return collect(Folder::getFilesByType($path, 'yaml'))->map(function ($file) {
+            try {
+                $data = YAML::parse(File::get($file));
+            } catch (ParseException $e) {
+                $data = [];
+                Log::warning('Could not parse form submission file: '.$file);
+            }
+
             return $this->makeSubmission()
                 ->id(pathinfo($file)['filename'])
-                ->data(YAML::parse(File::get($file)));
+                ->data($data);
         });
     }
